// src/app/interfaces/interfaces.ts
// Generated by https://quicktype.io

export interface RespuestaMDB {
        page:          number;
        total_results: number;
        total_pages:   number;
        results:       Pelicula[];
}

export interface Pelicula {
        popularity:        number;
        vote_count:        number;
        video:             boolean;
        poster_path:       string;
        id:                number;
        adult:             boolean;
        backdrop_path:     string;
        original_language: OriginalLanguage;
        original_title:    string;
        genre_ids:         number[];
        title:             string;
        vote_average:      number;
        overview:          string;
        release_date:      string;
}

export enum OriginalLanguage {
        En = "en",
        Es = "es",
}

// Generated by https://quicktype.io

export interface PeliculaDetalle {
        adult?:                 boolean;
        backdrop_path?:         string;
        belongs_to_collection?: null;
        budget?:                number;
        genres?:                Genre[];
        homepage?:              string;
        id?:                    number;
        imdb_id?:               string;
        original_language?:     string;
        original_title?:        string;
        overview?:              string;
        popularity?:            number;
        poster_path?:           string;
        production_companies?:  ProductionCompany[];
        production_countries?:  ProductionCountry[];
        release_date?:          string;
        revenue?:               number;
        runtime?:               number;
        spoken_languages?:      SpokenLanguage[];
        status?:                string;
        tagline?:               string;
        title?:                 string;
        video?:                 boolean;
        vote_average?:          number;
        vote_count?:            number;
}

export interface Genre {
        id:   number;
        name: string;
}

export interface ProductionCompany {
        id:             number;
        logo_path:      null;
        name:           string;
        origin_country: string;
}

export interface ProductionCountry {
        iso_3166_1: string;
        name:       string;
}

export interface SpokenLanguage {
        iso_639_1: string;
        name:      string;
}


// Generated by https://quicktype.io

export interface RespuestaCredits {
        id:   number;
        cast: Cast[];
        crew: Crew[];
}

export interface Cast {
        cast_id:      number;
        character:    string;
        credit_id:    string;
        gender:       number;
        id:           number;
        name:         string;
        order:        number;
        profile_path: null | string;
}

export interface Crew {
        credit_id:    string;
        department:   string;
        gender:       number;
        id:           number;
        job:          string;
        name:         string;
        profile_path: null | string;
}

export interface FavoritoGenero {
        genero:string,
        peliculas:PeliculaDetalle[]
}